# [aarch64] vectorize tolower/toupper with NEON