The patch uses a table lookup to cache the result of tolower/toupper function
calls.  When comparing against the sequential version on calling tolower/toupper
on a 1024 bytes string, this patch is 1553% faster (16x speedup) on a Graviton
A1 instance.
```
---------------------------------------------------------
Benchmark               Time             CPU   Iterations
---------------------------------------------------------
toupper_neon          411 ns          411 ns      1701697
toupper_scalar       6796 ns         6795 ns       103050
```
The following benchmark executes in 5.147s on master and 2.490s with the patch
a 2x speedup on an A1 Graviton instance:
```
function bmk() {
    $a = "Test";
    for ($i = 0; $i < 50000; $i++) {
        if (strtolower($a) === strtolower(strtoupper($a)))
            $a .= chr($i%256);
        else {
            echo "error at i=".$i."\n";
            echo "a=".$a."\n";
            echo "tolower(a)=".tolower($a)."\n";
            echo "tolower(toupper(a))=".tolower(toupper($a))."\n";
            return;
        }
    }
}
```
Tested with `make test` and the C code tested for all supported locales on my
ubuntu-18.04 aarch64-linux system.