@cmb69 2021-12-28
What is the status here? The patch has merge conflicts, and could at least be simplified due to https://wiki.php.net/rfc/strtolower-ascii.